Minor Updates
Added additional checking for "matrixType" in accordance to the phone application. There are 4 types of matrices that will be supported. The Updates Regarding this are located in  "LedEffects.cpp" on the "getPixelIndex(int x, int y)" function. In addition to this, I went ahead and removed some of the unnecessary printfs'.